requirements.txt
Django==3.0.3
requests_oauthlib==1.3.0
pyyaml==5.2
python-dateutil==2.8.1
xlwt==1.3.0
